made admin table responsive, delivery picking orders, cart payments affect profit

+++ server.js
@@ -26,32 +26,53 @@ app.post('/webhook', express.raw({ type: 'application/json' }), async (req, res)
       expand: ['line_items']
     });
 
-    const userId = session.metadata.user_id;
+    const monthIndex = new Date().getMonth()
+    const monthNames = [
+      "january", "february", "march", "april", "may", "june",
+      "july", "august", "september", "october", "november", "december"
+    ]
+    let PaymentType = session.metadata.paying_for
 
-    let userCart = (await axios.get(`https://dailymart-5c550-default-rtdb.firebaseio.com/cart/${userId}.json`)).data
+    if (PaymentType == 'cart') {
+      const userId = session.metadata.user_id;
 
-    const orderData = {
-      items: userCart,
-      total: session.amount_total / 100,
-      status: 'Processing',
-      createdAt: new Date().toISOString(),
-      customerName: session.metadata.customer_name,
-      customerEmail: session.customer_email
-    };
+      let userCart = (await axios.get(`https://dailymart-5c550-default-rtdb.firebaseio.com/cart/${userId}.json`)).data
 
-    try {
-      // Add the order to Firebase under the user's orders
-      const response = await axios.post(`https://dailymart-5c550-default-rtdb.firebaseio.com/orders/${userId}.json`, orderData);
-      if (response.status === 200) {
-        console.log('Order added successfully to Firebase');
-        await axios.delete(`https://dailymart-5c550-default-rtdb.firebaseio.com/cart/${userId}.json`);
-      } else {
-        console.error('Failed to add order to Firebase');
+      const orderData = {
+        items: userCart,
+        total: session.amount_total / 100,
+        status: 'Processing',
+        createdAt: new Date().toISOString(),
+        customerName: session.metadata.customer_name,
+        customerEmail: session.customer_email
+      };
+
+      try {
+        const response = await axios.post(`https://dailymart-5c550-default-rtdb.firebaseio.com/orders/${userId}.json`, orderData);
+        if (response.status === 200) {
+          await axios.delete(`https://dailymart-5c550-default-rtdb.firebaseio.com/cart/${userId}.json`);
+          let oldSalesRevenue = (await axios.get(`https://dailymart-5c550-default-rtdb.firebaseio.com/profits/${monthNames[monthIndex]}/salesRevenue.json`)).data
+
+          await axios.patch(`https://dailymart-5c550-default-rtdb.firebaseio.com/profits/${monthNames[monthIndex]}.json`, { salesRevenue: oldSalesRevenue + session.amount_total / 100 })
+        } else {
+          console.error('Failed to add order to Firebase');
+        }
+      } catch (error) {
+        console.error('Error adding order to Firebase:', error);
       }
-    } catch (error) {
-      console.error('Error adding order to Firebase:', error);
+    }
+
+    else {
+      try {
+        let oldSubscriptionsRevenue = (await axios.get(`https://dailymart-5c550-default-rtdb.firebaseio.com/profits/${monthNames[monthIndex]}/subscriptionsRevenue.json`)).data
+
+        await axios.patch(`https://dailymart-5c550-default-rtdb.firebaseio.com/profits/${monthNames[monthIndex]}.json`, { subscriptionsRevenue: oldSubscriptionsRevenue + 250 })
+      } catch (error) {
+        console.error('Error Patching the subscritpion revenue:', error);
+      }
     }
   }
+
 
   res.json({ received: true });
 });
@@ -101,6 +122,7 @@ app.post('/create-checkout-session', async (req, res) => {
       metadata: {
         customer_name: userName,
         user_id: userId,
+        paying_for: "cart"
         // cart: JSON.stringify(cartArray)
       }
     });
